Update Toolchain Source action to export QEMU source
Add support to map the QEMU component to 'qemu' in the Toolchain Source
module code.
Remove the 'Packages' component from Source Release testruns for both
MIPS and nanoMIPS, this is not exported it just wastes time checking it
out and then discarding it.
Update the version of the Toolchain Source action used for nanoMIPS to
pull in the QEMU source as well.